+++ src/Home/oauth.js
@@ -1,0 +1,32 @@
+function oauthSignIn() {
+    // Google's OAuth 2.0 endpoint for requesting an access token
+    var oauth2Endpoint = 'https://thepc-one.herokuapp.com/api/google';
+  
+    // Create <form> element to submit parameters to OAuth 2.0 endpoint.
+    var form = document.createElement('form');
+    form.setAttribute('method', 'GET'); // Send as a GET request.
+    form.setAttribute('action', oauth2Endpoint);
+  
+    // Parameters to pass to OAuth 2.0 endpoint.
+    var params = {'client_id': '390060085294-k1l5r25ugf2jpsqorsmns7m8o3ject6f.apps.googleusercontent.com',
+                  'redirect_uri': 'http://localhost:3000/home',
+                  'response_type': 'token',
+                  // 'scope': 'https://www.googleapis.com/auth/drive.metadata.readonly',
+                  // 'include_granted_scopes': 'true',
+                  'state': 'pass-through value'};
+  
+    // Add form parameters as hidden input values.
+    for (var p in params) {
+      var input = document.createElement('input');
+      input.setAttribute('type', 'hidden');
+      input.setAttribute('name', p);
+      input.setAttribute('value', params[p]);
+      form.appendChild(input);
+       }
+
+      // Add form to page and submit it to open the OAuth 2.0 endpoint.
+      document.body.appendChild(form);
+      form.submit();
+     }
+
+     export default oauthSignIn
